refactor: add stable id to scheduler assignment schema, assign id in scheduler service
