+++ apps/web/src/components/dashboard/team/schedule/schedule-grid.tsx
@@ -11,7 +11,7 @@ import { ScheduleDerivedData } from '~/hooks/schedule/useScheduleDerivedData'
 import { ScheduleFilters } from '~/hooks/schedule/useScheduleFilters'
 import { useScheduleStore } from '~/store/schedule.store'
 import { ScheduleRow } from './schedule-row'
-import { ScheduleTeamMemberHeaderCell } from './team-member-header-cell'
+import { ScheduleTeamMemberHeaderCell } from './schedule-team-member-header-cell'
 
 interface ScheduleGridProps {
   data: ScheduleData
@@ -54,7 +54,7 @@ export const ScheduleGrid = ({
           <div
             id={day.id}
             key={day.id}
-            className='sticky top-0 z-30 border-b border-r border-input flex items-center bg-background py-2 px-4 gap-1.5'
+            className='sticky top-0 z-30 border-b not-last:border-r border-input flex items-center bg-background py-2 px-4 gap-1.5'
           >
             <span className='font-bold'>
               {format(day.date, 'ccc', { locale })}
