+++ packages/api/src/schemas/schedule.ts
@@ -1,15 +1,8 @@
+import { SchedulerAssignmentSchema } from '@fuku/domain/schemas'
 import * as z from 'zod/v4'
 
-export const SchedulerAssignmentSchema = z.object({
-  teamMemberId: z.string(),
-  shiftTypeId: z.string(),
-  date: z.coerce.date(),
-  score: z.number().optional(),
-})
-
-export type SchedulerAssignment = z.infer<typeof SchedulerAssignmentSchema>
-
 export const GenerateScheduleOutputSchema = z.object({
+  id: z.string(),
   teamId: z.string(),
   period: z.object({
     start: z.date(),
