+++ packages/domain/src/schemas/index.ts
@@ -12,6 +12,7 @@ export * from './payGrade'
 export * from './payGradeShiftType'
 export * from './rule'
 export * from './ruleCondition'
+export * from './schedulerAssignment'
 export * from './shiftAssignment'
 export * from './shiftType'
 export * from './staffingRequirement'
